feat(ranking): refactor ranking data to use mock data from separate file
